feat: add group control in event details
