+++ src/app/features/create-event/event-details-form/event-details-form.component.html
@@ -15,6 +15,12 @@
       </mat-error>
     </mat-form-field>
   </div>
+  <mat-form-field *ngIf="isGroup" appearance="fill">
+    <mat-label>Wybierz grupę</mat-label>
+    <mat-select formControlName="group">
+      <mat-option [value]="1">1</mat-option>
+    </mat-select>
+  </mat-form-field>
   <div class="toggle">
     <mat-label class="toggle__label">Potwierdzenie obowiązkowe</mat-label>
     <mat-slide-toggle color="primary" formControlName="isConfirmationRequired"></mat-slide-toggle>
